backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
python-multipart==0.0.6
pydantic==2.5.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
httpx==0.25.2
lxml==4.9.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
PyJWT==2.8.0
